pandora/sphinx/requirements.txt
Flask
Flask-Bootstrap
Flask-HTTPAuth
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-PageDown
Flask-SQLAlchemy
Flask-Cache
Flask-Script
Flask-WTF
Jinja2
Mako
Markdown
MarkupSafe
PyMySQL
SQLAlchemy
WTForms
Werkzeug
alembic
bleach
blinker
html5lib
itsdangerous
six
Pillow
simplejson

ForgeryPy
Pygments
colorama
coverage
httpie
requests
selenium
ipython